fix(plugins): split channel hook context types
