+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -1,2 +1,2 @@
-048386e82c1f522442dcae2447dcee0cff809d3dfde7f66e42413acf7894d658  plugin-sdk-api-baseline.json
-cc3cedb20757f33b620841b675c401a82b2f5aa855ebf02f9b5ae93e505fb4dc  plugin-sdk-api-baseline.jsonl
+da3373338b7f9c5f5639ad8233a32897d2346a0babe69a77386a7bff154cdcb1  plugin-sdk-api-baseline.json
+17404d885e0d64ebc8e3c99443921058a8f1aebf76a5e612eb1f0cd7817d48f0  plugin-sdk-api-baseline.jsonl
