+++ src/plugins/hook-channel-context.types.ts
@@ -1,0 +1,18 @@
+export interface PluginHookChannelSenderContext {
+  /** Channel-scoped sender ID, matching `ctx.senderId` when both are present. */
+  id?: string;
+  [key: string]: unknown;
+}
+
+export interface PluginHookChannelChatContext {
+  /** Transport-native conversation ID, matching `ctx.chatId` when both are present. */
+  id?: string;
+  [key: string]: unknown;
+}
+
+export interface PluginHookChannelContext {
+  /** Sender metadata supplied by the originating channel. */
+  sender?: PluginHookChannelSenderContext;
+  /** Chat/conversation metadata supplied by the originating channel. */
+  chat?: PluginHookChannelChatContext;
+}
